refactor: ProductResponse로 응답을 분리
