+++ src/main/java/cart/web/controller/product/ProductViewController.java
@@ -2,7 +2,6 @@ package cart.web.controller.product;
 
 import cart.domain.product.ProductCategory;
 import cart.domain.product.ProductService;
-import cart.web.controller.product.dto.ProductRequest;
 import cart.web.controller.product.dto.ProductResponse;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -20,25 +19,11 @@ public class ProductViewController {
         this.productService = productService;
     }
 
-    @GetMapping("/")
-    public String renderIndex(final Model model) {
-        final List<ProductResponse> products = productService.getProducts();
-        model.addAttribute("products", products);
-        return "index.html";
-    }
-
-    @GetMapping("/admin")
-    public String renderAdmin(final Model model) {
-        final List<ProductResponse> products = productService.getProducts();
-        model.addAttribute("products", products);
-        return "admin.html";
-    }
-
     // TODO: 관리자의 권한으로 접근하면, 소비자가 접근할 때와 다른 화면을 보여줄 필요가 있지 않을까
     @GetMapping("/products/{productId}")
     public String renderProduct(@PathVariable Long productId, final Model model) {
-        final ProductRequest productRequest = productService.getById(productId);
-        model.addAttribute("product", productRequest);
+        final ProductResponse productResponse = productService.getById(productId);
+        model.addAttribute("product", productResponse);
         return "product.html";
     }
