+++ src/main/java/cart/web/controller/user/UserViewController.java
@@ -1,6 +1,8 @@
 package cart.web.controller.user;
 
+import cart.domain.product.ProductService;
 import cart.domain.user.UserService;
+import cart.web.controller.product.dto.ProductResponse;
 import cart.web.controller.user.dto.UserResponse;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -12,9 +14,19 @@ import java.util.List;
 public class UserViewController {
 
     private final UserService userService;
+    private final ProductService productService;
 
-    public UserViewController(final UserService userService) {
+    public UserViewController(final UserService userService, final ProductService productService) {
         this.userService = userService;
+        this.productService = productService;
+    }
+
+
+    @GetMapping("/")
+    public String renderIndex(final Model model) {
+        final List<ProductResponse> products = productService.getProducts();
+        model.addAttribute("products", products);
+        return "index.html";
     }
 
     @GetMapping("/settings")
